Add static calculation feature to matcalc/relaxation.py
In the original relaxation.py, users are allowed to do cell relaxation or atoms only relaxation. Static calculation is an important feature itself and also part of some complex applications (eg. QHA). As MatCalc provides a universal interface for different uMLIPs, how about we add this feature to RelaxCalc? The default is still cell relaxation, and user can set "relax_atoms" and "relax_cell" to False to do static calculation. 

Signed-off-by: Runze Liu <146490083+rul048@users.noreply.github.com>